fix for win32
git-svn-id: http://svn.cern.ch/guest/lhcb/Gauss/trunk@93136 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 7faaca19e539ac09e32454c662549c699f33643c

+++ Gauss/Sim/GaussTools/src/Lib/MCTruthManager.cpp
@@ -1,4 +1,4 @@
-// $Id: MCTruthManager.cpp,v 1.5 2006-05-02 18:56:03 gcorti Exp $
+// $Id: MCTruthManager.cpp,v 1.6 2007-10-02 16:33:25 gcorti Exp $
 // Include files 
 
 // local
@@ -259,7 +259,7 @@ void MCTruthManager::printTree(HepMC::GenParticle* particle, std::string offset)
   {
     std::string deltaoffset = "";
 
-    if( fmod((*it)->barcode(),10000000) != fmod(particle->barcode(),10000000) )
+    if( fmod(double((*it)->barcode()),10000000) != fmod(double(particle->barcode()),10000000) )
       {
         deltaoffset = " | ";
       }
